#15 Fixes "Swift runtime failure: arithmetic overflow" error in Tests

+++ Tests/Swift/CARBPerformanceTests.swift
@@ -26,7 +26,7 @@ class CARBSwiftPerformanceTests: XCTestCase {
          var status: RingBufferError
          for iteration in 0 ..< numberOfIterations {
             status = ringBuffer.store(writeBuffer.audioBufferList, framesToWrite: Int64(IOCapacity),
-                                      startWrite: Int64(IOCapacity * iteration))
+                                      startWrite: Int64(IOCapacity) * Int64(iteration))
             if status != .noError {
                fatalError()
             }
